requirements.txt
openai==0.27.8
langchain==0.0.202
tiktoken==0.4.0
pandas==2.0.2
yfinance==0.2.20
python-dotenv==1.0.0